+++ winforge/__init__.py
@@ -2,5 +2,5 @@
 WinForge System Optimization Package.
 """
 
-__version__ = "1.0.5"
+__version__ = "1.0.6"
 __author__ = "0xdowz"